+++ examples/alb_test_fixture/variables.tf
@@ -1,0 +1,11 @@
+variable "log_location_prefix" {
+  default = "my-lb-logs"
+}
+
+variable "region" {
+  default = "us-west-2"
+}
+
+variable "log_bucket_name" {
+  default = "test-log-bucket"
+}