+++ examples/test_fixtures/data.tf
@@ -1,5 +1,0 @@
-data "aws_caller_identity" "current" {}
-
-data "aws_availability_zones" "available" {}
-
-data "aws_region" "current" {}