+++ examples/test_fixtures/variables.tf
@@ -1,7 +1,0 @@
-variable "log_location_prefix" {
-  default = "my-alb-logs"
-}
-
-variable "region" {
-  default = "us-west-2"
-}