+++ main.tf
@@ -1,84 +1,86 @@
-resource "aws_alb" "main" {
-  name            = "${var.alb_name}"
-  subnets         = ["${var.subnets}"]
-  security_groups = ["${var.alb_security_groups}"]
-  internal        = "${var.alb_is_internal}"
-  tags            = "${merge(var.tags, map("Name", var.alb_name))}"
+locals {
+  prefix_length = "${length(var.load_balancer_name) < 6 ? length(var.load_balancer_name) : 6 }"
+  name_prefix   = "${substr(var.load_balancer_name, 0, local.prefix_length)}"
+}
+
+resource "aws_lb" "application" {
+  load_balancer_type         = "application"
+  name_prefix                = "${local.name_prefix}"
+  internal                   = "${var.load_balancer_is_internal}"
+  security_groups            = ["${var.load_balancer_security_groups}"]
+  subnets                    = ["${var.subnets}"]
+  idle_timeout               = "${var.idle_timeout}"
+  enable_deletion_protection = "${var.enable_deletion_protection}"
+  enable_http2               = "${var.enable_http2}"
+  ip_address_type            = "${var.ip_address_type}"
+  tags                       = "${merge(var.tags, map("Name", var.load_balancer_name))}"
 
   access_logs {
+    enabled = true
     bucket  = "${var.log_bucket_name}"
     prefix  = "${var.log_location_prefix}"
-    enabled = "${var.enable_logging}"
   }
 
-  depends_on = ["aws_s3_bucket.log_bucket"]
+  timeouts {
+    create = "${var.load_balancer_create_timeout}"
+    delete = "${var.load_balancer_delete_timeout}"
+    update = "${var.load_balancer_update_timeout}"
+  }
 }
 
-resource "aws_s3_bucket" "log_bucket" {
-  bucket        = "${var.log_bucket_name}"
-  policy        = "${var.bucket_policy == "" ? data.aws_iam_policy_document.bucket_policy.json : var.bucket_policy}"
-  force_destroy = "${var.force_destroy_log_bucket}"
-  count         = "${var.create_log_bucket ? 1 : 0}"
-  tags          = "${merge(var.tags, map("Name", var.log_bucket_name))}"
-}
-
-resource "aws_alb_target_group" "target_group" {
-  name                 = "${var.alb_name}-tg"
-  port                 = "${var.backend_port}"
-  protocol             = "${upper(var.backend_protocol)}"
+resource "aws_lb_target_group" "main" {
+  name                 = "${lookup(var.target_groups[count.index], "name")}"
   vpc_id               = "${var.vpc_id}"
-  deregistration_delay = "${var.deregistration_delay}"
+  port                 = "${lookup(var.target_groups[count.index], "backend_port")}"
+  protocol             = "${upper(lookup(var.target_groups[count.index], "backend_protocol"))}"
+  deregistration_delay = "${lookup(var.target_groups[count.index], "deregistration_delay", lookup(var.target_groups_defaults, "deregistration_delay"))}"
+  target_type          = "${lookup(var.target_groups[count.index], "target_type", lookup(var.target_groups_defaults, "target_type"))}"
 
   health_check {
-    interval            = "${var.health_check_interval}"
-    path                = "${var.health_check_path}"
-    port                = "${var.health_check_port}"
-    healthy_threshold   = "${var.health_check_healthy_threshold}"
-    unhealthy_threshold = "${var.health_check_unhealthy_threshold}"
-    timeout             = "${var.health_check_timeout}"
-    protocol            = "${var.backend_protocol}"
-    matcher             = "${var.health_check_matcher}"
+    interval            = "${lookup(var.target_groups[count.index], "health_check_interval", lookup(var.target_groups_defaults, "health_check_interval"))}"
+    path                = "${lookup(var.target_groups[count.index], "health_check_path", lookup(var.target_groups_defaults, "health_check_path"))}"
+    port                = "${lookup(var.target_groups[count.index], "health_check_port", lookup(var.target_groups_defaults, "health_check_port"))}"
+    healthy_threshold   = "${lookup(var.target_groups[count.index], "health_check_healthy_threshold", lookup(var.target_groups_defaults, "health_check_healthy_threshold"))}"
+    unhealthy_threshold = "${lookup(var.target_groups[count.index], "health_check_unhealthy_threshold", lookup(var.target_groups_defaults, "health_check_unhealthy_threshold"))}"
+    timeout             = "${lookup(var.target_groups[count.index], "health_check_timeout", lookup(var.target_groups_defaults, "health_check_timeout"))}"
+    protocol            = "${upper(lookup(var.target_groups[count.index], "healthcheck_protocol", lookup(var.target_groups[count.index], "backend_protocol")))}"
+    matcher             = "${lookup(var.target_groups[count.index], "health_check_matcher", lookup(var.target_groups_defaults, "health_check_matcher"))}"
   }
-
-  target_type = "${var.target_type}"
 
   stickiness {
     type            = "lb_cookie"
-    cookie_duration = "${var.cookie_duration}"
-    enabled         = "${ var.cookie_duration == 1 ? false : true}"
+    cookie_duration = "${lookup(var.target_groups[count.index], "cookie_duration", lookup(var.target_groups_defaults, "cookie_duration"))}"
+    enabled         = "${lookup(var.target_groups[count.index], "stickiness_enabled", lookup(var.target_groups_defaults, "stickiness_enabled"))}"
   }
 
-  tags = "${merge(var.tags, map("Name", "${var.alb_name}-tg"))}"
+  tags  = "${merge(var.tags, map("Name", lookup(var.target_groups[count.index], "name")))}"
+  count = "${var.target_groups_count}"
 
-  depends_on = ["aws_alb.main"]
+  depends_on = ["aws_lb.application"]
 }
 
-resource "aws_alb_listener" "frontend_http" {
-  load_balancer_arn = "${aws_alb.main.arn}"
-  port              = "${var.alb_http_port}"
-  protocol          = "HTTP"
-  count             = "${contains(var.alb_protocols, "HTTP") ? 1 : 0}"
+resource "aws_lb_listener" "frontend_http_tcp" {
+  load_balancer_arn = "${aws_lb.application.arn}"
+  port              = "${lookup(var.http_tcp_listeners[count.index], "port")}"
+  protocol          = "${lookup(var.http_tcp_listeners[count.index], "protocol")}"
+  count             = "${var.http_tcp_listeners_count}"
 
   default_action {
-    target_group_arn = "${aws_alb_target_group.target_group.id}"
+    target_group_arn = "${aws_lb_target_group.main.*.id[lookup(var.http_tcp_listeners[count.index], "target_group_index", 0)]}"
     type             = "forward"
   }
-
-  depends_on = ["aws_alb.main"]
 }
 
-resource "aws_alb_listener" "frontend_https" {
-  load_balancer_arn = "${aws_alb.main.arn}"
-  port              = "${var.alb_https_port}"
+resource "aws_lb_listener" "frontend_https" {
+  load_balancer_arn = "${aws_lb.application.arn}"
+  port              = "${lookup(var.https_listeners[count.index], "port")}"
   protocol          = "HTTPS"
-  certificate_arn   = "${var.certificate_arn}"
-  ssl_policy        = "${var.security_policy}"
-  count             = "${contains(var.alb_protocols, "HTTPS") ? 1 : 0}"
+  certificate_arn   = "${lookup(var.https_listeners[count.index], "certificate_arn")}"
+  ssl_policy        = "${lookup(var.https_listeners[count.index], "ssl_policy", var.listener_ssl_policy_default)}"
+  count             = "${var.https_listeners_count}"
 
   default_action {
-    target_group_arn = "${aws_alb_target_group.target_group.id}"
+    target_group_arn = "${aws_lb_target_group.main.*.id[lookup(var.https_listeners[count.index], "target_group_index", 0)]}"
     type             = "forward"
   }
-
-  depends_on = ["aws_alb.main"]
 }