+++ test/integration/default/test_alb.rb
@@ -3,40 +3,27 @@
 require 'awspec'
 require 'rhcl'
 
-module_vars = Rhcl.parse(File.open('examples/test_fixtures/variables.tf'))
+module_vars = Rhcl.parse(File.open('examples/alb_test_fixture/variables.tf'))
 log_location_prefix = module_vars['variable']['log_location_prefix']['default']
 
 # rubocop:disable LineLength
 state_file = 'terraform.tfstate.d/kitchen-terraform-default-aws/terraform.tfstate'
 tf_state = JSON.parse(File.open(state_file).read)
-principal_account_id = tf_state['modules'][0]['outputs']['principal_account_id']['value']
+http_tcp_listener_arns = tf_state['modules'][0]['outputs']['http_tcp_listener_arns']['value']
+https_listener_arns = tf_state['modules'][0]['outputs']['https_listener_arns']['value']
+target_group_arns = tf_state['modules'][0]['outputs']['target_group_arns']['value']
+# rubocop:enable LineLength
+alb_arn = tf_state['modules'][0]['outputs']['alb_id']['value']
+alb_name = alb_arn.split('/')[-2]
 account_id = tf_state['modules'][0]['outputs']['account_id']['value']
 region = tf_state['modules'][0]['outputs']['region']['value']
 ENV['AWS_REGION'] = region
-# this must match the format in examples/test_fixtures/locals.tf
-log_bucket_name = 'logs-' + region + '-' + account_id
-policy = "{
-    \"Version\": \"2012-10-17\",
-    \"Statement\": [
-        {
-            \"Sid\": \"AllowToPutLoadBalancerLogsToS3Bucket\",
-            \"Effect\": \"Allow\",
-            \"Principal\": {
-                \"AWS\": \"arn:aws:iam::#{principal_account_id}:root\"
-            },
-            \"Action\": \"s3:PutObject\",
-            \"Resource\": \"arn:aws:s3:::#{log_bucket_name}/#{log_location_prefix}/AWSLogs/#{account_id}/*\"
-        }
-    ]
-}"
-# rubocop:enable LineLength
-log_object = "#{log_location_prefix}/AWSLogs/#{account_id}/ELBAccessLogTestFile"
 vpc_id = tf_state['modules'][0]['outputs']['vpc_id']['value']
 security_group_id = tf_state['modules'][0]['outputs']['sg_id']['value']
 
-describe alb('test-alb') do
+describe alb(alb_name) do
   it { should exist }
-  its(:load_balancer_name) { should eq 'test-alb' }
+  its(:load_balancer_name) { should eq alb_name }
   its(:vpc_id) { should eq vpc_id }
   it { should belong_to_vpc('test-vpc') }
   its(:type) { should eq 'application' }
@@ -45,17 +32,44 @@ describe alb('test-alb') do
   it { should have_security_group(security_group_id) }
 end
 
-describe alb_target_group('test-alb-tg') do
-  it { should exist }
-  its(:health_check_path) { should eq '/' }
-  its(:health_check_port) { should eq 'traffic-port' }
-  its(:health_check_protocol) { should eq 'HTTP' }
-  it { should belong_to_alb('test-alb') }
-  it { should belong_to_vpc('test-vpc') }
+describe alb_target_group('foo') do
+  its(:port) { should eq 80 }
 end
 
-describe s3_bucket(log_bucket_name) do
-  it { should exist }
-  it { should have_object(log_object) }
-  it { should have_policy(policy) }
+describe alb_target_group('bar') do
+  its(:port) { should eq 8080 }
+end
+
+target_group_arns.each do |tg_arn|
+  tg_name = tg_arn.split('/')[-2]
+  describe alb_target_group(tg_name) do
+    it { should exist }
+    it { should belong_to_alb(alb_name) }
+    it { should belong_to_vpc('test-vpc') }
+    its(:protocol) { should eq 'HTTP' }
+    its(:health_check_protocol) { should eq 'HTTP' }
+    its(:health_check_interval_seconds) { should eq 10 }
+    its(:health_check_timeout_seconds) { should eq 5 }
+    its(:healthy_threshold_count) { should eq 3 }
+    its(:health_check_path) { should eq '/' }
+    its(:unhealthy_threshold_count) { should eq 3 }
+    its(:target_type) { should eq 'instance' }
+    its(:health_check_port) { should eq 'traffic-port' }
+  end
+end
+
+https_listener_arns.each do |listener|
+  describe alb_listener(listener) do
+    it { should exist }
+    its(:protocol) { should eq 'HTTPS' }
+    its(:port) { should eq(443).or(eq(8443)) }
+  end
+end
+
+http_tcp_listener_arns.each do |listener|
+  describe alb_listener(listener) do
+    it { should exist }
+    its(:protocol) { should eq 'HTTP' }
+    its(:port) { should eq(80).or(eq(8080).or(eq(8081))) }
+  end
 end